+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Seqera Platform deep-link button now resolves `watchUrl` from the `observers` list used by Nextflow 24.10, in addition to `observersV1` and `observersV2` ([#73](https://github.com/seqeralabs/nf-slack/pull/73))
+
 ## [0.5.1] - 2026-02-19
 
 ### Fixed